The taste pile, less the two that were not improvements
Case folding was written twice, once as a loop inside
`location_names_are_unique` and once as a lambda inside
`same_ignoring_case`. Two spellings of one rule is two chances to
disagree about it, so there is one `fold_case` above both.

The identity fill in `operand_for_parameter` is `std::ranges::iota`.

`Name::length` is a `std::uint8_t`. The capacity is 15, and `Resolved`
carries two of these as a template argument on a translation unit that
peaks above a gigabyte.

`read_verb` claimed two things depended on the spelling, the concept and
the scan. Only the scan does; the concept cannot, since it does not know
what a location is. `write` needs no constant, being spelled once at the
call the generator splices, and the comment now says so.

The two index narrowings that had no limit have one. A reference holds
its vocabulary as a byte and a goto its target the same way, so 256 of
either is a capacity like the eight members of a vocabulary, and now says
so rather than wrapping.

Two left alone. `parse_pattern`'s loop wants the index to derive a bit
position, and `views::enumerate` yields a signed difference type, so the
ranges version trades a raw loop for a cast. And `PatternTest`'s four
`CHECK_THROWS` would need a trait to detect "not a constant expression"
before they could be `STATIC_CHECK`s, which is more machinery than the
inconsistency costs.

+++ z80/v4/refract/include/refract/Model.hpp
@@ -25,7 +25,10 @@ namespace specbolt::refract {
 // reason; this one is v4's.
 struct Name {
   std::array<char, 15> storage{};
-  std::size_t length{};
+  // A byte, not a `std::size_t`: `Resolved` is a template argument on a
+  // translation unit that peaks above a gigabyte, so the eight bytes are worth
+  // not spending.
+  std::uint8_t length{};
   constexpr Name() = default;
   template<std::size_t N>
   constexpr Name(const char (&text)[N]) : Name(std::string_view{text, N - 1}) {} // NOLINT(*-explicit-constructor)
@@ -33,7 +36,7 @@ struct Name {
     if (text.size() > storage.size())
       throw std::length_error("name does not fit");
     std::ranges::copy(text, storage.begin());
-    length = text.size();
+    length = static_cast<std::uint8_t>(text.size());
   }
   static constexpr std::size_t capacity = std::tuple_size_v<decltype(storage)>;
   [[nodiscard]] constexpr std::string_view view() const { return {storage.data(), length}; }
